requirements.txt
cffi==1.1.2
cryptography==0.9.3
Django==1.8.1
django-classy-tags==0.6.1
django-countries==3.3
django-crispy-forms==1.4.0
django-dotenv==1.3.0
django-email-as-username==1.6.11
django-getenv==1.3.1
django-sekizai==0.8.1
easy-thumbnails==2.2
enum34==1.0.4
idna==2.0
ipaddress==1.0.11
paypalrestsdk==1.10.0
Pillow==2.9.0
pyasn1==0.1.8
pycparser==2.14
pyOpenSSL==0.15.1
requests==2.7.0
six==1.9.0
wheel==0.24.0
